Refactor the code in ExecutorPositionsUtil
