+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/common/utils/CommonUtil.java
@@ -1256,77 +1256,6 @@ public class CommonUtil {
     }
 
     /**
-     * Checks if the provided location (interpreted relatively to the provided module) is location in the same file
-     * provided.
-     *
-     * @param module   Module where the location resides
-     * @param location Location
-     * @param filePath File path to check against
-     * @return True if the location resides in the provided file path
-     * @throws IOException On IO errors
-     */
-    public static boolean isLocationInFile(Module module, Location location, Path filePath) throws IOException {
-        Path symbolPath;
-        if (module.project().kind() == ProjectKind.SINGLE_FILE_PROJECT) {
-            symbolPath = module.project().sourceRoot();
-        } else if (module.isDefaultModule()) {
-            symbolPath = module.project().sourceRoot().resolve(location.lineRange().filePath());
-        } else {
-            symbolPath = module.project().sourceRoot()
-                    .resolve("modules")
-                    .resolve(module.moduleName().moduleNamePart())
-                    .resolve(filePath);
-        }
-
-        return Files.isSameFile(symbolPath, filePath);
-    }
-
-    /**
-     * Check if the symbol is a class symbol with self as the name.
-     *
-     * @param symbol               Symbol
-     * @param context              PositionedOperationContext
-     * @param enclosedModuleMember ModuleMemberDeclarationNode
-     * @return {@link Boolean} whether the symbol is a self class symbol.
-     */
-    public static boolean isSelfClassSymbol(Symbol symbol, PositionedOperationContext context,
-                                            @Nonnull ModuleMemberDeclarationNode enclosedModuleMember) {
-        Optional<String> name = symbol.getName();
-        if (enclosedModuleMember.kind() != SyntaxKind.CLASS_DEFINITION || symbol.kind() != SymbolKind.VARIABLE
-                || name.isEmpty() || !name.get().equals(SELF_KW)) {
-            return false;
-        }
-
-        Optional<Symbol> memberSymbol = context.workspace().semanticModel(context.filePath())
-                .flatMap(semanticModel -> semanticModel.symbol(enclosedModuleMember));
-
-        if (memberSymbol.isEmpty() || memberSymbol.get().kind() != SymbolKind.CLASS) {
-            return false;
-        }
-        ClassSymbol classSymbol = (ClassSymbol) memberSymbol.get();
-        VariableSymbol selfSymbol = (VariableSymbol) symbol;
-        TypeSymbol varTypeSymbol = CommonUtil.getRawType(selfSymbol.typeDescriptor());
-
-        return classSymbol.equals(varTypeSymbol);
-    }
-
-    /**
-     * Check if the symbol is an object symbol with self as the name.
-     *
-     * @param symbol       Symbol
-     * @param nodeAtCursor Node
-     * @return {@link Boolean} whether the symbol is a self object symbol.
-     */
-    public static boolean isSelfObjectSymbol(Symbol symbol, Node nodeAtCursor) {
-        Node currentNode = nodeAtCursor;
-        while (currentNode != null && currentNode.kind() != SyntaxKind.OBJECT_CONSTRUCTOR) {
-            currentNode = currentNode.parent();
-        }
-        return currentNode != null && currentNode.kind() == SyntaxKind.OBJECT_CONSTRUCTOR
-                && symbol.getName().orElse("").equals(SELF_KW);
-    }
-
-    /**
      * Check if the cursor is positioned in a lock statement node context.
      *
      * @param context Completion context.
